Remove stale required tasks lists
EC will choose require task list with the newest effective on date
older than the current time, so the older lists are no longer
relevant.

In theory if EC was run with an effective time in the past these
lists might become effective, but that's not something we do in
practice.